Thu, Sep 7, 2023, 1:00 PM +01:00

+++ llamas.c
@@ -4,7 +4,7 @@
 
 int main(void)
 {
-    float startPo;
+    int startPo;
     int targetPo;
     float unrounded;
